Restructure Token Parsing
Restrucutre the token parsing code in the discovery client to avoid
smashing the stack.  Newer versions of GCC generate a runtime error
when smashing the stack.  This happened in the token parsing code
because it was running past the end of the array looking for tokens.

Replaced the custom tokenization code with strsep(3) that is made
for this sort of operation.